adapt to windows
To build the demo properly, it was necessary to make a static wxVTK library instead of shared.
The way events are recorded based on the mouse position on the render window seems different between macOS and windows. Accordingly I adapted the SetEventInformationCentralize function and removed the SetEventInformationFromOrigin function.

Would need to check if it works on macOS and linux as it is or if two/three environments need to be defined with different event behaviors.