Fix issue with ButtonUp event
Previously, renderer would get stuck in zoom-mode if the control key was
released before the ButtonUp event (fired by releasing the mouse
button). This was caused by the ButtonUp event handler returning when
the EventType didn't match. I'm not sure why this check was ever there
in the first place.